fix: service enablement fixes and manual tests

+++ generators/service/languages/node-express/index.js
@@ -13,7 +13,6 @@ const GENERATOR_LOCATION = 'server';
 const PATH_MAPPINGS_FILE = "./server/config/mappings.json";
 const PATH_LOCALDEV_CONFIG_FILE = "server/localdev-config.json";
 const PATH_PACKAGE_JSON = "./package.json";
-const PATH_GIT_IGNORE = "./.gitignore";
 
 module.exports = class extends Generator {
 	constructor(args, opts) {
@@ -30,32 +29,11 @@ module.exports = class extends Generator {
 		this.context.addDependencies = this._addDependencies.bind(this);
 		this.context.addMappings = this._addMappings.bind(this);
 		this.context.addLocalDevConfig = this._addLocalDevConfig.bind(this);
-	}
 
-	writing() {
 		let serviceCredentials,
 			scaffolderKey,
 			serviceKey;
 		this._addDependencies(this.fs.read(this.templatePath() + "/" + this.context.dependenciesFile));
-
-		
-		this.fs.copy(
-			this.templatePath() + "/service-manager.js",
-			this.destinationPath("./server/services/service-manager.js")
-		);
-
-		this.fs.copy(
-			this.templatePath() + "/services-index.js",
-			this.destinationPath("./server/services/index.js")
-		);
-		
-		// Add PATH_LOCALDEV_CONFIG_FILE to .gitignore
-		let gitIgnorePath = this.destinationPath(PATH_GIT_IGNORE);
-		if (this.fs.exists(gitIgnorePath)){
-			this.fs.append(gitIgnorePath, PATH_LOCALDEV_CONFIG_FILE);
-		} else {
-			this.fs.write(gitIgnorePath, PATH_LOCALDEV_CONFIG_FILE);
-		}
 
 		//initializing ourselves by composing with the service generators
 		let root = path.dirname(require.resolve('../..'));
@@ -78,6 +56,18 @@ module.exports = class extends Generator {
 		});
 	}
 
+	writing() {
+		// this.fs.copy(
+		// 	this.templatePath() + "/service-manager.js",
+		// 	this.destinationPath("./server/services/service-manager.js")
+		// );
+
+		// this.fs.copy(
+		// 	this.templatePath() + "/services-index.js",
+		// 	this.destinationPath("./server/services/index.js")
+		// );
+	}
+
 	_addDependencies(serviceDepdendenciesString){
 		let serviceDependencies = JSON.parse(serviceDepdendenciesString);
 		let packageJsonPath = this.destinationPath(PATH_PACKAGE_JSON);
@@ -97,9 +87,10 @@ module.exports = class extends Generator {
 	end(){
 		// Remove GENERATE_HERE from /server/services/index.js
 		let servicesIndexJsFilePath = this.destinationPath("./server/services/index.js");
-		let indexFileContent = this.fs.read(servicesIndexJsFilePath);
-		indexFileContent = indexFileContent.replace(GENERATE_HERE, "");
-		this.fs.write(servicesIndexJsFilePath, indexFileContent);
-	  
+		if (this.fs.exists(servicesIndexJsFilePath)) {
+			let indexFileContent = this.fs.read(servicesIndexJsFilePath);
+			indexFileContent = indexFileContent.replace(GENERATE_HERE, "");
+			this.fs.write(servicesIndexJsFilePath, indexFileContent);
+		}
 	}
 };
